refactor(agents): GENNY_CONFIGS registry; BudgetConfig is ValidatedConfig
Reconcile the recipe foundation with dev (post Genny-rename #394):
- genny_swe_config.py -> genny_configs.py, adding a GENNY_CONFIGS registry
  built from the existing make_agent_config (no duplicated config logic).
- BudgetConfig now subclasses ValidatedConfig so nested tweaks
  (agent.budget.cost_limit = 2.0) validate at the assignment site.
